Run the animation timer on the main thread to fix interaction bug
Run the animation timer on the main thread.  We want to allow the user
to interact with the UI while this timer is running.  If we run it on
this thread, the timer will be halted while the user is touching the
screen (that's why the chart was disappearing in our collection view).